Combine Requestable and Reviewable
The functionality is the same for all requestables, and now that we've
added the necessary columns to all the requestable models, we can
combine the two concerns.
