+++ app/models/concerns/requestable.rb
@@ -13,6 +13,7 @@ module Requestable
     validates :state, presence: true, inclusion: { in: states.values }
 
     validates :received_at, presence: true, if: :received?
+    validates :reviewed_at, presence: true, unless: -> { passed.nil? }
 
     define_method :received! do
       update!(state: "received", received_at: Time.zone.now)
@@ -23,6 +24,21 @@ module Requestable
 
   def expired_at
     created_at + expires_after
+  end
+
+  def reviewed!(passed)
+    update!(passed:, reviewed_at: Time.zone.now)
+  end
+
+  def failed
+    return nil if passed.nil?
+    passed == false
+  end
+
+  def status
+    return state if passed.nil?
+
+    passed ? "accepted" : "rejected"
   end
 
   def after_received(user:)
