+++ app/models/further_information_request.rb
@@ -31,10 +31,6 @@ class FurtherInformationRequest < ApplicationRecord
 
   FOUR_WEEK_COUNTRY_CODES = %w[AU CA GI NZ US].freeze
 
-  def failed
-    passed == false
-  end
-
   def expires_after
     if !application_form.created_under_new_regulations? &&
          FOUR_WEEK_COUNTRY_CODES.include?(application_form.country.code)
