+++ spec/view_objects/assessor_interface/further_information_request_view_object_spec.rb
@@ -85,7 +85,7 @@ RSpec.describe AssessorInterface::FurtherInformationRequestViewObject do
 
     context "when not passed and not recommended" do
       before do
-        further_information_request.update!(passed: nil)
+        further_information_request.reviewed!(true)
         assessment.request_further_information!
       end
       it { is_expected.to be true }
@@ -93,7 +93,7 @@ RSpec.describe AssessorInterface::FurtherInformationRequestViewObject do
 
     context "when passed and not recommended" do
       before do
-        further_information_request.update!(passed: true)
+        further_information_request.reviewed!(true)
         assessment.request_further_information!
       end
       it { is_expected.to be true }
@@ -109,7 +109,7 @@ RSpec.describe AssessorInterface::FurtherInformationRequestViewObject do
 
     context "when passed and recommended" do
       before do
-        further_information_request.update!(passed: true)
+        further_information_request.reviewed!(true)
         assessment.award!
       end
       it { is_expected.to be false }
